Do not try to remove tmp path that does not exist
If the extension check mode is set to “On file change” this will crash the
extension.

+++ Scripts/provider.js
@@ -108,7 +108,9 @@ class IssuesProvider {
 
             console.info("Found " + parser.issues.length + " issue(s)");
 
-            nova.fs.remove(tmpPath);
+            if (tmpPath) {
+                nova.fs.remove(tmpPath);
+            }
             resolve(parser.issues);
             parser.clear();
         });
